English
English is an international language which is the most popular language around the world. By learning English, you can find many opportunities as job, education and living abroad.
English belongs to the Indo-European family of languages and is therefore a related to most other languages spoken in Europe and western Asia from Iceland to India. 
Modern English is analytic, whereas proto-Indo_ European, the ancestral tongue of most of the modern European languages (German, French, Russian, Greek), was synthetic or inflicted.
English has official or special status in at least 75 countries with a combined population of two billion people. It is estimated that one out of four people worldwide speak English with some degree of competence.
Many countries teach and encourage youngsters to acquire English as a second language. Even in nations where English is not as official language, many science and engineering curriculum are written in English.
English abilities will most certainly aid you in any business endeavors you choose to pursue. Many large corporations will only hire professional employees after determining whether or not they speak good English.
The English language opens an ocean of career opportunities to those who speak this language anywhere in the world. Similarly, or has turned into and inevitable requirement for various fields and professions like medicine, computing and more.
Tin the fast- evolving world, it is essential to have a common language is clear now, we move on to why we must learn the English language. first of all, it is a global language. It is so common that one out of five people can speak or understand this language.

There are many sources for learning English language as books, courses, online classes, sites, social media... one of the best sites for learning English for non-native English speakers, is zabanman.com. This site is operated in august 25th 2022 by specialist team and now has a complete package of English studies.
Zabanman.com released an English grammar book which is written in Persian language. As long as you want to have that book, refer to the site and download the book.
Good luck
Author: Sohrab Rostami